requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
anndata==0.9.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backports.zoneinfo;python_version<"3.9"
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
deprecation==2.1.0
Django==4.2.2
django-elasticsearch-dsl==7.3
django-email-verification==0.3.1
django-social-share==2.3.0
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
frozenlist==1.4.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.7.0
Jinja2==3.1.2
kombu==5.3.1
llvmlite==0.40.1rc1
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.1.1
natsort==8.3.1
numba==0.57.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
prompt-toolkit==3.0.39
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
rabbitmq==0.2.0
requests==2.31.0
scipy==1.10.1
share==1.0.4
six==1.16.0
social==0.1.dev3
sqlparse==0.4.4
stripe==5.5.0
tomli==2.0.1
tqdm==4.65.0
twilio==8.9.0
typing-extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
zipp==3.15.0
